Move dependencies to top level
