+++ includeBuild/dependencies/src/main/kotlin/depsKapt.kt
@@ -1,0 +1,31 @@
+/*
+*  Copyright 2021 Matthew Nelson
+*
+*  Licensed under the Apache License, Version 2.0 (the "License");
+*  you may not use this file except in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+*  Unless required by applicable law or agreed to in writing, software
+*  distributed under the License is distributed on an "AS IS" BASIS,
+*  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+*  See the License for the specific language governing permissions and
+*  limitations under the License.
+* */
+@file:Suppress("SpellCheckingInspection", "unused", "ClassName")
+
+/**
+ * Kapt dependencies
+ * */
+object depsKapt {
+
+    object google {
+        const val hilt                      = "com.google.dagger:hilt-compiler:${versions.google.hilt}"
+    }
+
+    object square {
+        const val moshi                     = "com.squareup.moshi:moshi-kotlin-codegen:${versions.square.moshi}"
+    }
+
+}
